backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-openai
chromadb
pydantic
requests
python-multipart
langchain-groq
langchain-huggingface
sentence-transformers
gunicorn
fpdf2
pytest
httpx
groq
moviepy
gradio_client
python-pptx
